flask-cv/requirements.txt
click==8.1.3
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
opencv-python==4.5.5.62
Pillow==9.5.0
python-dotenv==1.0.0
Werkzeug==2.2.3
